Feature add search

+++ src/utils/convertDate.js
@@ -1,7 +1,7 @@
 const convertDate = (date) => {
   const properDate = new Date(date.split("T")[0]);
   const day = properDate.getDate();
-  const month = properDate.toLocaleString("default", { month: "long" });
+  const month = properDate.toLocaleString("default", { month: "short" });
   const year = properDate.getFullYear();
   const UTCDate = month + " " + day + ", " + year;
   return UTCDate;
